+++ docs/Whats-New.md
@@ -150,6 +150,7 @@ New:
 - Add developer command to dump all existing triggers, tags, and local and global variables to the log output (by Rampastring)
 - Fix a bug where aircraft are unable to attack shrouded targets in campaign games and instead get stuck in mid-air (by Rampastring)
 - Fix a bug where the player was able to input keyboard commands while input was locked through a trigger action (by Rampastring)
+- Fix a bug where a vehicle transport could end up attached to its own cargo, causing the transport to disappear upon unloading (by Rampastring)
 
 Vanilla fixes:
 - Fix HouseType `Nod` having the `Prefix=B` and `Side=GDI` in vanilla `rules.ini` by setting them to `N` and `Nod`, respectively (by CCHyper/tomsons26)